backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
supabase==2.6.0
gunicorn==21.2.0
pandas==2.2.2
openpyxl==3.1.5

